fixed bug when core db does not have contigs
Removed assumption that the 3 levels of assembly are chromosome,
supercontig and contig, as there are many other species with different
coord_system sequence level (e.g. primary assembly, scaffold).

A similar error might be raised in the future regarding supercontigs,
and since 'toplevel' and 'seqlevel' are linked in the assembly table,
that part of the code has also been removed.