Add prototype Gitlab-CI configuration
What I have done here is take .gitlab-ci.yml and
MultiTestDB.conf.gitlabci.mysql from ensembl (my own development version
in case of the former), modify the before_script section of the former
to account for differences in dependencies between the two projects,
create a dependent-build trigger attempting to launch one on
ensembl-gh-mirror/ensembl (i.e.  already on EBI Gitlab), and fix the
declaration of same because it has turned out variable expansion
in trigger declarations works only in the branch field, not in "project".

Note that all of this results in a somewhat different build
matrix than on Travis but unlike there, on GitLab we have got
consistency between ensembl and ensembl-test.